wrap dashboard rapper to get fetch dashboards
