+++ lib/groups/screens/group_home.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:provider/provider.dart';
 import 'package:receipt_wrangler_mobile/groups/widgets/group_app_bar.dart';
-import 'package:receipt_wrangler_mobile/groups/widgets/group_dashboard.dart';
+import 'package:receipt_wrangler_mobile/groups/widgets/group_dashboard_wrapper.dart';
 import 'package:receipt_wrangler_mobile/groups/widgets/groups_bottom_nav.dart';
 import 'package:receipt_wrangler_mobile/models/group_model.dart';
 import 'package:receipt_wrangler_mobile/shared/widgets/screen_wrapper.dart';
@@ -32,7 +32,7 @@ class _GroupHome extends State<GroupHome> {
             },
           )),
       bottomNavigationBarWidget: const GroupsBottomNav(),
-      children: const [GroupDashboard()],
+      children: const [GroupDashboardWrapper()],
     );
   }
 }
